removed old request flow and updated
