+++ src/navigation/HomeStack/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import WelcomeScreen from '../../screens/WelcomeScreen';
 import SetupScreen from '../../screens/SetupScreen';
 import LocationSetupScreen from '../../screens/LocationSetupScreen';
-import RequestFlow from '../../screens/RequestFlow';
 import LoginScreen from '../../screens/LoginScreen';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
@@ -10,7 +9,6 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 type HomeStackParamList = {
   Welcome: any;
   LocationSetup: any;
-  RequestFlow: any;
   Setup: any;
   Login: any;
 };
@@ -28,11 +26,6 @@ const HomeStack = () => {
           name="LocationSetup"
           options={{ headerShown: false }}
           component={LocationSetupScreen}
-        />
-        <Stack.Screen
-          options={{ headerShown: false }}
-          name="RequestFlow"
-          component={RequestFlow}
         />
         <Stack.Screen
           options={{ headerShown: false }}
