+++ src/navigation/WalletStack/index.tsx
@@ -11,7 +11,7 @@ type WalletStackParamList = {
 const Stack = createNativeStackNavigator<WalletStackParamList>();
 const WalletStack = () => {
   return (
-    <NavigationContainer>
+    <NavigationContainer  independent={true}>
       <Stack.Navigator initialRouteName="Wallet">
         <Stack.Screen
           component={WalletScreen}
