+++ src/screens/MainCardScreen/index.tsx
@@ -1,4 +1,5 @@
 export { default as WalletScreen } from './WalletScreen';
+export { default as WalletRequestScreen } from './WalletRequestScreen';
 export { default as ExchangeScreen } from './ExchangeScreen';
 export { default as HomeScreen } from './HomeScreen';
 export { default as SettingScreen } from './SettingScreen';
